# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a blog API.",
        "title": "Blog API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/blog/posts": {
            "get": {
                "description": "全てのブログ記事のリストを取得します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blog"
                ],
                "summary": "ブログ記事リストを取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.BlogPost"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "新しいブログ記事を作成します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blog"
                ],
                "summary": "新しいブログ記事を作成",
                "parameters": [
                    {
                        "description": "ブログ記事",
                        "name": "blogPost",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BlogPost"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.BlogPost"
                        }
                    }
                }
            }
        },
        "/api/v1/blog/posts/{id}": {
            "get": {
                "description": "指定されたIDのブログ記事を取得します。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blog"
                ],
                "summary": "特定のブログ記事を取得",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "記事ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BlogPost"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.BlogPost": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}